feature: sge testing utilizing an actual sge host
