+++ bbi_sge_test.go
@@ -2,13 +2,10 @@ package babylontest
 
 import (
 	"context"
-	"github.com/ghodss/yaml"
-	"github.com/metrumresearchgroup/babylon/configlib"
+	"github.com/metrumresearchgroup/gogridengine"
 	log "github.com/sirupsen/logrus"
-	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"testing"
 	"time"
@@ -40,8 +37,6 @@ func TestBabylonCompletesSGEExecution(t *testing.T){
 		//log.Infof("Beginning SGE execution test for model set %s",v.identifier)
 		v.Prepare(ctx)
 
-		bbiBinary, _ := exec.LookPath("bbi")
-
 		for _ , m := range v.models {
 
 			nonMemArguments := []string{
@@ -51,8 +46,6 @@ func TestBabylonCompletesSGEExecution(t *testing.T){
 				"sge",
 				"--nm_version",
 				os.Getenv("NMVERSION"),
-				"--babylon_binary",
-				bbiBinary,
 			}
 
 			_, err := m.Execute(v,nonMemArguments...)
@@ -61,15 +54,7 @@ func TestBabylonCompletesSGEExecution(t *testing.T){
 				t.Error(err)
 			}
 
-
-			os.Chdir(filepath.Join(v.Workpath,m.identifier))
-			//Now let's run the script that was generated
-			_, err = executeCommand(ctx,filepath.Join(v.Workpath,m.identifier,"grid.sh"))
-			os.Chdir(whereami)
-
-			if err != nil {
-				log.Error(err)
-			}
+			WaitForSGEToTerminate(v)
 
 			testingDetails := NonMemTestingDetails{
 				t:         t,
@@ -113,8 +98,6 @@ func TestBabylonCompletesParallelSGEExecution(t *testing.T){
 		//log.Infof("Beginning SGE parallel execution test for model set %s",v.identifier)
 		v.Prepare(ctx)
 
-		bbiBinary, _ := exec.LookPath("bbi")
-
 		for _ , m := range v.models {
 
 			nonMemArguments := []string{
@@ -124,8 +107,6 @@ func TestBabylonCompletesParallelSGEExecution(t *testing.T){
 				"sge",
 				"--nm_version",
 				os.Getenv("NMVERSION"),
-				"--babylon_binary",
-				bbiBinary,
 				"--parallel=true",
 				"--mpi_exec_path",
 				os.Getenv("MPIEXEC_PATH"),
@@ -162,82 +143,7 @@ func TestBabylonCompletesParallelSGEExecution(t *testing.T){
 	}
 }
 
-func TestConfigValuesAreCorrectInWrittenFile(t *testing.T){
-	//Get BB and make sure we have the test data moved over.
-	//Clean Slate
 
-	const qsub string = "/usr/local/bin/qsub"
-	purgeBinary(qsub)
-	fakeBinary(qsub)
-
-	//Pick a few critical configuration components such as
-	/*
-		--clean_level 3
-		--copy_level 1
-		--debug
-		--parallel=true <- make sure it's present
-		--mpi_exec_path
-	 */
-
-	Scenario := InitializeScenarios([]string{
-		"240",
-	})[0]
-
-	Scenario.Prepare(context.Background())
-
-	commandAndArgs := []string{
-		"--debug=true", //Needs to be in debug mode to generate the expected output
-		"nonmem",
-		"run",
-		"--clean_lvl",
-		"3",
-		"--copy_lvl",
-		"1",
-		"--parallel=true",
-		"--mpi_exec_path",
-		os.Getenv("MPIEXEC_PATH"),
-		"local",
-		"--nm_version",
-		os.Getenv("NMVERSION"),
-	}
-
-	for _, m := range Scenario.models {
-		output, err := m.Execute(Scenario,commandAndArgs...)
-
-		assert.Nil(t,err)
-		assert.NotEmpty(t,output)
-
-		nmd := NonMemTestingDetails{
-			t:         t,
-			OutputDir: filepath.Join(Scenario.Workpath,m.identifier),
-			Model:     m,
-			Output:    output,
-		}
-
-		AssertNonMemCompleted(nmd)
-		AssertNonMemCreatedOutputFiles(nmd)
-		AssertNonMemOutputContainsParafile(nmd)
-
-		//Now read the Config Lib
-		configFile := filepath.Join(Scenario.Workpath,m.identifier,"babylon.yaml")
-		file, _ := os.Open(configFile)
-		Config := configlib.Config{}
-		bytes, _ := ioutil.ReadAll(file)
-		err = yaml.Unmarshal(bytes,&Config)
-
-		assert.Nil(t,err)
-
-		assert.Equal(t,3,Config.CleanLvl)
-		assert.Equal(t,1,Config.CopyLvl)
-		assert.Equal(t, true,Config.Parallel,)
-		assert.Equal(t,os.Getenv("NMVERSION"), Config.NMVersion)
-
-		assert.Equal(t,os.Getenv("MPIEXEC_PATH"),Config.MPIExecPath )
-		assert.Equal(t,false,Config.Overwrite)
-	}
-
-	purgeBinary(qsub)
-}
 
 
 
@@ -252,4 +158,22 @@ func fakeBinary(name string) {
 
 func purgeBinary(name string) {
 	os.Remove(name)
+}
+
+
+func WaitForSGEToTerminate(scenario *Scenario) {
+	for CountOfPendingJobs() > 0 {
+		log.Infof("Located %d pending jobs. Waiting for 30 seconds to check again", CountOfPendingJobs())
+		time.Sleep(30 * time.Second)
+	}
+
+	log.Info("Looks like all queued and running jobs have terminated")
+}
+
+func CountOfPendingJobs() int {
+	jobs, _ := gogridengine.GetJobsWithFilter(func(j gogridengine.Job) bool {
+		return j.State == "qw" || j.State == "r"
+	})
+
+	return len(jobs)
 }
