+++ config_test.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"github.com/metrumresearchgroup/babylon/cmd"
+	"github.com/metrumresearchgroup/babylon/configlib"
 	"github.com/stretchr/testify/assert"
+	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -58,4 +60,76 @@ func AssertBBIConfigContainsSpecifiedNMVersion(details NonMemTestingDetails, nmV
 	assert.NotNil(details.t,nm)
 	assert.NotEqual(details.t,nm,cmd.NonMemModel{})
 	assert.Equal(details.t,nm.Configuration.NMVersion,nmVersion)
+}
+
+
+func TestConfigValuesAreCorrectInWrittenFile(t *testing.T){
+	//Get BB and make sure we have the test data moved over.
+	//Clean Slate
+	//Pick a few critical configuration components such as
+	/*
+		--clean_level 3
+		--copy_level 1
+		--debug
+		--parallel=true <- make sure it's present
+		--mpi_exec_path
+	*/
+
+	Scenario := InitializeScenarios([]string{
+		"240",
+	})[0]
+
+	Scenario.Prepare(context.Background())
+
+	commandAndArgs := []string{
+		"--debug=true", //Needs to be in debug mode to generate the expected output
+		"nonmem",
+		"run",
+		"--clean_lvl",
+		"3",
+		"--copy_lvl",
+		"1",
+		"--parallel=true",
+		"--mpi_exec_path",
+		os.Getenv("MPIEXEC_PATH"),
+		"local",
+		"--nm_version",
+		os.Getenv("NMVERSION"),
+	}
+
+	for _, m := range Scenario.models {
+		output, err := m.Execute(Scenario,commandAndArgs...)
+
+		assert.Nil(t,err)
+		assert.NotEmpty(t,output)
+
+		nmd := NonMemTestingDetails{
+			t:         t,
+			OutputDir: filepath.Join(Scenario.Workpath,m.identifier),
+			Model:     m,
+			Output:    output,
+		}
+
+		AssertNonMemCompleted(nmd)
+		AssertNonMemCreatedOutputFiles(nmd)
+		AssertNonMemOutputContainsParafile(nmd)
+
+		//Now read the Config Lib
+		configFile := filepath.Join(Scenario.Workpath,m.identifier,"babylon.yaml")
+		file, _ := os.Open(configFile)
+		Config := configlib.Config{}
+		bytes, _ := ioutil.ReadAll(file)
+		err = yaml.Unmarshal(bytes,&Config)
+
+		assert.Nil(t,err)
+
+		assert.Equal(t,3,Config.CleanLvl)
+		assert.Equal(t,1,Config.CopyLvl)
+		assert.Equal(t, true,Config.Parallel,)
+		assert.Equal(t,os.Getenv("NMVERSION"), Config.NMVersion)
+
+		assert.Equal(t,os.Getenv("MPIEXEC_PATH"),Config.MPIExecPath )
+		assert.Equal(t,false,Config.Overwrite)
+	}
+
 }
