+++ nmqual_test.go
@@ -2,11 +2,14 @@ package babylontest
 
 import (
 	"context"
+	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
+	"github.com/metrumresearchgroup/gogridengine"
+	"time"
 )
 
 func TestNMQUALExecutionSucceeds(t *testing.T){
@@ -59,3 +62,5 @@ func AssertScriptContainsAutologReference(details NonMemTestingDetails){
 
 	assert.Contains(details.t,scriptFileContent,"autolog.pl")
 }
+
+
